+++ sdk/highlight-run/src/client/index.tsx
@@ -118,7 +118,7 @@ import {
 } from '@opentelemetry/api'
 import { IntegrationClient } from '../integrations'
 import { LaunchDarklyIntegration } from '../integrations/launchdarkly'
-import { LDClientMin } from '../integrations/launchdarkly/types/LDClient'
+import { LDClient } from '../integrations/launchdarkly'
 import { createLog, defaultLogOptions } from './listeners/console-listener'
 import { CustomSampler } from './otel/sampling/CustomSampler'
 
@@ -1566,7 +1566,7 @@ SessionSecureID: ${this.sessionData.sessionSecureID}`,
 		this._lastSnapshotTime = new Date().getTime()
 	}
 
-	registerLD(client: LDClientMin) {
+	registerLD(client: LDClient) {
 		// TODO(vkorolik): can only register one LD client for now
 		if (this._integrations.length) return
 		this._integrations.push(new LaunchDarklyIntegration(client))
@@ -1603,7 +1603,7 @@ export type {
 	HighlightOptions,
 	HighlightPublicInterface,
 	Integration,
-	LDClientMin,
+	LDClient,
 	Metadata,
 	Metric,
 	MixpanelIntegrationOptions,