+++ sdk/highlight-run/src/client/listeners/network-listener/performance-listener.ts
@@ -1,4 +1,4 @@
-type NetworkInformation = {
+export type NetworkInformation = {
 	downlink: number
 	downlinkMax: number
 	rtt: number