+++ sdk/@launchdarkly/observability/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @launchdarkly/observability
 
+## 0.1.5
+
+### Patch Changes
+
+- 24f38cf: update launchdarkly package reporting of attributes for application and environment
+
+## 0.1.4
+
+### Patch Changes
+
+- 96d5818: change type exports to be compatible with moduleResolution node
+- Updated dependencies [96d5818]
+    - highlight.run@9.18.1
+
+## 0.1.3
+
+### Patch Changes
+
+- 6e77363: include highlight.run as dependency for type resolution
+
+## 0.1.2
+
+### Patch Changes
+
+- 1ea87fd: correctly publish dist directory
+
+## 0.1.1
+
+### Patch Changes
+
+- bcac87b: release initial public version
+
 ## 0.1.0
 
 ### Minor Changes